hw_final/ml_service/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
pydantic
starlette-exporter
python-json-logger
psycopg2-binary
